Clear up import policy
