+++ docs/source/telegram.error.rst
@@ -5,5 +5,4 @@ telegram.error module
 
 .. automodule:: telegram.error
     :members:
-    :undoc-members:
     :show-inheritance:
