+++ docs/source/telegram.helpers.rst
@@ -1,0 +1,8 @@
+:github_url: https://github.com/python-telegram-bot/python-telegram-bot/blob/master/telegram/helpers.py
+
+telegram.helpers Module
+=======================
+
+.. automodule:: telegram.helpers
+    :members:
+    :show-inheritance:
