+++ docs/source/telegram.request.rst
@@ -1,0 +1,8 @@
+:github_url: https://github.com/python-telegram-bot/python-telegram-bot/blob/master/telegram/request.py
+
+telegram.request
+================
+
+.. automodule:: telegram.request
+    :members:
+    :show-inheritance:
