--- docs/source/telegram.utils.request.rst
@@ -1,8 +1,0 @@
-:github_url: https://github.com/python-telegram-bot/python-telegram-bot/blob/master/telegram/utils/request.py
-
-telegram.utils.request.Request
-==============================
-
-.. autoclass:: telegram.utils.request.Request
-    :members:
-    :show-inheritance:
