+++ examples/inlinebot.py
@@ -16,8 +16,8 @@ import logging
 from uuid import uuid4
 
 from telegram import InlineQueryResultArticle, ParseMode, InputTextMessageContent, Update
+from telegram.helpers import escape_markdown
 from telegram.ext import Updater, InlineQueryHandler, CommandHandler, CallbackContext
-from telegram.utils.helpers import escape_markdown
 
 # Enable logging
 logging.basicConfig(
