+++ telegram/__init__.py
@@ -56,7 +56,6 @@ from .replymarkup import ReplyMarkup
 from .replykeyboardmarkup import ReplyKeyboardMarkup
 from .replykeyboardremove import ReplyKeyboardRemove
 from .forcereply import ForceReply
-from .error import TelegramError, PassportDecryptionError
 from .files.inputfile import InputFile
 from .files.file import File
 from .parsemode import ParseMode
@@ -130,16 +129,6 @@ from .files.inputmedia import (
     InputMediaAnimation,
     InputMediaAudio,
     InputMediaDocument,
-)
-from .constants import (
-    MAX_MESSAGE_LENGTH,
-    MAX_CAPTION_LENGTH,
-    SUPPORTED_WEBHOOK_PORTS,
-    MAX_FILESIZE_DOWNLOAD,
-    MAX_FILESIZE_UPLOAD,
-    MAX_MESSAGES_PER_SECOND_PER_CHAT,
-    MAX_MESSAGES_PER_SECOND,
-    MAX_MESSAGES_PER_MINUTE_PER_GROUP,
 )
 from .passport.passportelementerrors import (
     PassportElementError,
@@ -261,13 +250,6 @@ __all__ = (  # Keep this alphabetically ordered
     'LabeledPrice',
     'Location',
     'LoginUrl',
-    'MAX_CAPTION_LENGTH',
-    'MAX_FILESIZE_DOWNLOAD',
-    'MAX_FILESIZE_UPLOAD',
-    'MAX_MESSAGES_PER_MINUTE_PER_GROUP',
-    'MAX_MESSAGES_PER_SECOND',
-    'MAX_MESSAGES_PER_SECOND_PER_CHAT',
-    'MAX_MESSAGE_LENGTH',
     'MaskPosition',
     'Message',
     'MessageAutoDeleteTimerChanged',
@@ -298,7 +280,6 @@ __all__ = (  # Keep this alphabetically ordered
     'ReplyKeyboardRemove',
     'ReplyMarkup',
     'ResidentialAddress',
-    'SUPPORTED_WEBHOOK_PORTS',
     'SecureData',
     'SecureValue',
     'ShippingAddress',
@@ -307,8 +288,6 @@ __all__ = (  # Keep this alphabetically ordered
     'Sticker',
     'StickerSet',
     'SuccessfulPayment',
-    'PassportDecryptionError',
-    'TelegramError',
     'TelegramObject',
     'Update',
     'User',
