+++ telegram/bot.py
@@ -100,7 +100,7 @@ from telegram.utils.helpers import (
     parse_file_input,
     DEFAULT_20,
 )
-from telegram.utils.request import Request
+from telegram.request import Request
 from telegram.utils.types import FileInput, JSONDict, ODVInput, DVInput
 
 if TYPE_CHECKING:
@@ -157,8 +157,8 @@ class Bot(TelegramObject):
         token (:obj:`str`): Bot's unique authentication.
         base_url (:obj:`str`, optional): Telegram Bot API service URL.
         base_file_url (:obj:`str`, optional): Telegram Bot API file URL.
-        request (:obj:`telegram.utils.request.Request`, optional): Pre initialized
-            :obj:`telegram.utils.request.Request`.
+        request (:obj:`telegram.request.Request`, optional): Pre initialized
+            :obj:`telegram.request.Request`.
         private_key (:obj:`bytes`, optional): Private key for decryption of telegram passport data.
         private_key_password (:obj:`bytes`, optional): Password for above private key.
         defaults (:class:`telegram.ext.Defaults`, optional): An object containing default values to
@@ -313,7 +313,7 @@ class Bot(TelegramObject):
         if reply_markup is not None:
             if isinstance(reply_markup, ReplyMarkup):
                 # We need to_json() instead of to_dict() here, because reply_markups may be
-                # attached to media messages, which aren't json dumped by utils.request
+                # attached to media messages, which aren't json dumped by telegram.request
                 data['reply_markup'] = reply_markup.to_json()
             else:
                 data['reply_markup'] = reply_markup
@@ -4475,7 +4475,7 @@ class Bot(TelegramObject):
             data['contains_masks'] = contains_masks
         if mask_position is not None:
             # We need to_json() instead of to_dict() here, because we're sending a media
-            # message here, which isn't json dumped by utils.request
+            # message here, which isn't json dumped by telegram.request
             data['mask_position'] = mask_position.to_json()
 
         result = self._post('createNewStickerSet', data, timeout=timeout, api_kwargs=api_kwargs)
@@ -4555,7 +4555,7 @@ class Bot(TelegramObject):
             data['tgs_sticker'] = parse_file_input(tgs_sticker)
         if mask_position is not None:
             # We need to_json() instead of to_dict() here, because we're sending a media
-            # message here, which isn't json dumped by utils.request
+            # message here, which isn't json dumped by telegram.request
             data['mask_position'] = mask_position.to_json()
 
         result = self._post('addStickerToSet', data, timeout=timeout, api_kwargs=api_kwargs)
@@ -4877,7 +4877,7 @@ class Bot(TelegramObject):
         if reply_markup:
             if isinstance(reply_markup, ReplyMarkup):
                 # We need to_json() instead of to_dict() here, because reply_markups may be
-                # attached to media messages, which aren't json dumped by utils.request
+                # attached to media messages, which aren't json dumped by telegram.request
                 data['reply_markup'] = reply_markup.to_json()
             else:
                 data['reply_markup'] = reply_markup
@@ -5178,9 +5178,9 @@ class Bot(TelegramObject):
                 entities parsing. If not specified, the original caption is kept.
             parse_mode (:obj:`str`, optional): Mode for parsing entities in the new caption. See
                 the constants in :class:`telegram.ParseMode` for the available modes.
-            caption_entities (:class:`telegram.utils.types.SLT[MessageEntity]`): List of special
-                entities that appear in the new caption, which can be specified instead of
-                parse_mode
+            caption_entities (List[:class:`telegram.MessageEntity`], optional): List of special
+                entities that appear in the new caption, which can be specified instead
+                of parse_mode.
             disable_notification (:obj:`bool`, optional): Sends the message silently. Users will
                 receive a notification with no sound.
             reply_to_message_id (:obj:`int`, optional): If the message is a reply, ID of the
@@ -5219,7 +5219,7 @@ class Bot(TelegramObject):
         if reply_markup:
             if isinstance(reply_markup, ReplyMarkup):
                 # We need to_json() instead of to_dict() here, because reply_markups may be
-                # attached to media messages, which aren't json dumped by utils.request
+                # attached to media messages, which aren't json dumped by telegram.request
                 data['reply_markup'] = reply_markup.to_json()
             else:
                 data['reply_markup'] = reply_markup
