+++ telegram/ext/extbot.py
@@ -39,7 +39,7 @@ from ..utils.helpers import DEFAULT_NONE
 
 if TYPE_CHECKING:
     from telegram import InlineQueryResult, MessageEntity
-    from telegram.utils.request import Request
+    from telegram.request import Request
     from .defaults import Defaults
 
 HandledTypes = TypeVar('HandledTypes', bound=Union[Message, CallbackQuery, Chat])
