+++ telegram/user.py
@@ -22,12 +22,14 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Any, List, Optional, Union, Tuple
 
 from telegram import TelegramObject, constants
+from telegram.helpers import (
+    mention_markdown as helpers_mention_markdown,
+    mention_html as helpers_mention_html,
+)
 from telegram.utils.helpers import (
-    mention_html as util_mention_html,
     DEFAULT_NONE,
     DEFAULT_20,
 )
-from telegram.utils.helpers import mention_markdown as util_mention_markdown
 from telegram.utils.types import JSONDict, FileInput, ODVInput, DVInput
 
 if TYPE_CHECKING:
@@ -203,8 +205,8 @@ class User(TelegramObject):
 
         """
         if name:
-            return util_mention_markdown(self.id, name)
-        return util_mention_markdown(self.id, self.full_name)
+            return helpers_mention_markdown(self.id, name)
+        return helpers_mention_markdown(self.id, self.full_name)
 
     def mention_markdown_v2(self, name: str = None) -> str:
         """
@@ -216,8 +218,8 @@ class User(TelegramObject):
 
         """
         if name:
-            return util_mention_markdown(self.id, name, version=2)
-        return util_mention_markdown(self.id, self.full_name, version=2)
+            return helpers_mention_markdown(self.id, name, version=2)
+        return helpers_mention_markdown(self.id, self.full_name, version=2)
 
     def mention_html(self, name: str = None) -> str:
         """
@@ -229,8 +231,8 @@ class User(TelegramObject):
 
         """
         if name:
-            return util_mention_html(self.id, name)
-        return util_mention_html(self.id, self.full_name)
+            return helpers_mention_html(self.id, name)
+        return helpers_mention_html(self.id, self.full_name)
 
     def pin_message(
         self,
