+++ tests/test_audio.py
@@ -23,7 +23,7 @@ import pytest
 from flaky import flaky
 
 from telegram import Audio, TelegramError, Voice, MessageEntity, Bot
-from telegram.utils.helpers import escape_markdown
+from telegram.helpers import escape_markdown
 from tests.conftest import check_shortcut_call, check_shortcut_signature, check_defaults_handling
