+++ tests/test_document.py
@@ -24,7 +24,7 @@ from flaky import flaky
 
 from telegram import Document, PhotoSize, TelegramError, Voice, MessageEntity, Bot
 from telegram.error import BadRequest
-from telegram.utils.helpers import escape_markdown
+from telegram.helpers import escape_markdown
 from tests.conftest import check_shortcut_signature, check_shortcut_call, check_defaults_handling
