+++ tests/test_file.py
@@ -98,7 +98,7 @@ class TestFile:
         def test(*args, **kwargs):
             return self.file_content
 
-        monkeypatch.setattr('telegram.utils.request.Request.retrieve', test)
+        monkeypatch.setattr('telegram.request.Request.retrieve', test)
         out_file = file.download()
 
         try:
@@ -114,7 +114,7 @@ class TestFile:
         def test(*args, **kwargs):
             return self.file_content
 
-        monkeypatch.setattr('telegram.utils.request.Request.retrieve', test)
+        monkeypatch.setattr('telegram.request.Request.retrieve', test)
         file_handle, custom_path = mkstemp()
         try:
             out_file = file.download(custom_path)
@@ -144,7 +144,7 @@ class TestFile:
 
         file.file_path = None
 
-        monkeypatch.setattr('telegram.utils.request.Request.retrieve', test)
+        monkeypatch.setattr('telegram.request.Request.retrieve', test)
         out_file = file.download()
 
         assert out_file[-len(file.file_id) :] == file.file_id
@@ -158,7 +158,7 @@ class TestFile:
         def test(*args, **kwargs):
             return self.file_content
 
-        monkeypatch.setattr('telegram.utils.request.Request.retrieve', test)
+        monkeypatch.setattr('telegram.request.Request.retrieve', test)
         with TemporaryFile() as custom_fobj:
             out_fobj = file.download(out=custom_fobj)
             assert out_fobj is custom_fobj
@@ -178,7 +178,7 @@ class TestFile:
         def test(*args, **kwargs):
             return self.file_content
 
-        monkeypatch.setattr('telegram.utils.request.Request.retrieve', test)
+        monkeypatch.setattr('telegram.request.Request.retrieve', test)
 
         # Check that a download to a newly allocated bytearray works.
         buf = file.download_as_bytearray()
