+++ tests/test_photo.py
@@ -24,7 +24,7 @@ from flaky import flaky
 
 from telegram import Sticker, TelegramError, PhotoSize, InputFile, MessageEntity, Bot
 from telegram.error import BadRequest
-from telegram.utils.helpers import escape_markdown
+from telegram.helpers import escape_markdown
 from tests.conftest import (
     expect_bad_request,
     check_shortcut_call,
