+++ tests/test_video.py
@@ -24,7 +24,7 @@ from flaky import flaky
 
 from telegram import Video, TelegramError, Voice, PhotoSize, MessageEntity, Bot
 from telegram.error import BadRequest
-from telegram.utils.helpers import escape_markdown
+from telegram.helpers import escape_markdown
 from tests.conftest import check_shortcut_call, check_shortcut_signature, check_defaults_handling
